#177 moved validate.js into helpers directory
