+++ src/controllers/send.js
@@ -7,7 +7,7 @@ const email = sendemail.email;
 sendemail.set_template_directory('src/email_templates');
 const { getForm } = require('../model/form_queries');
 const { addData } = require('./helpers.js');
-const { validateSendEmail } = require('./validate');
+const { validateSendEmail } = require('./helpers/validate');
 
 exports.get = (req, res) => {
   res.render('send', {
